Add initial support for kubernetes 1.29
